requirements.txt
djangorestframework
django-cors-headers
# celery redis django-celery-beat